Use activeFedora current solr connection for the extract uri
Previously we were using the value of the yaml configuration, which is
the default way to initialize a connection, but not the only way. It is
also possible for the connection to be initialized programatically.
This is what hyku is doing to support multiple tenants each with their
own solr collection.

See https://github.com/samvera-labs/hyku/issues/1275