Fix shell execution
Closes stderr and stdout at the end of the process, rather than never
closing stderr, and closing stdout before there's time for any data to
come in.  This fixes a crash when processes (like opj_compress) print
out to stdout a lot.

Fixes an edge case where an EOFError causes an invocation of a
non-existent "Rails.logger" function.  In Rails, it would be
Rails.logger.debug or something, and outside Rails contexts, there is no
Rails.logger at all.  This was never noticed before, as EOFs weren't
detected because the previous logic never checked stdout, since it was
already closed.