rescue IO::EAGAINWaitReadable errors
Adds handling of IO::WaitReadable errors (such as IO::EAGAINWaitReadable) in the manner suggested by IO#read_nonblock documentation. The existing use of IO.select appears able to fail when only one of stdout or stderr is ready. Each file will now be retried if the read is not ready.