Update Dockerfile and installation instructions to reflect changes to Kakadu downloads.
Downloading the Kakadu demo now requires email verification, which causes a 404 error when running the wget command in the Dockerfile.

The new Dockerfile now copies the necesssary kakadu files from a .zip that the user downloads and places in the working directory. This commit also updates the instructions with this new step.